Move FlatTree implementation into a separate module.
This lets us write `from jax._src import flattree as ft` instead of
`from jax._src.tree_util import FlatTree as ft` where all the flattree
functions are static methods.
