+++ jax/_src/interpreters/remat.py
@@ -22,8 +22,9 @@ from jax._src import api_util
 from jax._src.util import (safe_map, safe_zip, unzip2, weakref_lru_cache,
                            partition_list)
 from jax._src.interpreters import partial_eval as pe
+from jax._src import flattree as ft
 from jax._src.tree_util import (
-    FlatTree, Partial, tree_unflatten, tree_leaves_checked)
+    Partial, tree_unflatten, tree_leaves_checked)
 from jax._src import source_info_util
 from jax._src.core import typeof
 
@@ -41,14 +42,14 @@ def remat_transform(policy, f, *args):
     jaxpr_trace = pe.DynamicJaxprTrace(None)
     jaxpr_trace.tag = core.TraceTag()
     trace = RematTrace(parent_trace, jaxpr_trace, policy)
-    args_ft = FlatTree.flatten_static_argnums_argnames(args, {}, (), ())
+    args_ft = ft.flatten_static_argnums_argnames(args, {}, (), ())
     in_tracers = args_ft.map(
         lambda x: RematTracer(trace, x, jaxpr_trace.new_arg(typeof(x), None)))  # type: ignore # noqa F821
     with core.set_current_trace(trace):
       args, kwargs = in_tracers.unflatten()
       ans_pytree = f(*args, **kwargs)
       dbg = dbg.set_result_paths(ans_pytree)
-      ans_ft = FlatTree.flatten(ans_pytree)
+      ans_ft = ft.flatten(ans_pytree)
       del ans_pytree, args, kwargs
     out_ft, out_tracer_ft = ans_ft.map(trace.to_val_tracer_pair).unzip2()
     src = source_info_util.current()
