+++ jax/_src/lax/control_flow/common.py
@@ -24,8 +24,8 @@ from jax._src import ad_util
 from jax._src import config
 from jax._src import core
 from jax._src.interpreters import partial_eval as pe
+from jax._src import flattree as ft
 from jax._src.tree_util import (
-    FlatTree,
     equality_errors_pytreedef,
     keystr,
     tree_map,
@@ -148,7 +148,7 @@ def _make_closed_jaxpr(
     debug_info: core.DebugInfo,
 ):
   closed_jaxpr, _ = pe.trace_to_jaxpr(
-      traceable, FlatTree.flatten_args(*in_avals), debug_info
+      traceable, ft.flatten_args(*in_avals), debug_info
   )
   return closed_jaxpr
