+++ jax/_src/lax/control_flow/solves.py
@@ -28,7 +28,8 @@ from jax._src.interpreters import batching
 from jax._src.interpreters import mlir
 from jax._src.interpreters import partial_eval as pe
 from jax._src.traceback_util import api_boundary
-from jax._src.tree_util import tree_leaves, FlatTree
+from jax._src import flattree as ft
+from jax._src.tree_util import tree_leaves
 from jax._src.util import split_list, safe_map
 import numpy as np
 
@@ -89,10 +90,10 @@ def custom_root(f: Callable,
     The result of calling solve(f, initial_guess) with gradients defined via
     implicit differentiation assuming ``f(solve(f, initial_guess)) == 0``.
   """
-  guess_flat = FlatTree.flatten(initial_guess)
+  guess_flat = ft.flatten(initial_guess)
   guess_avals = guess_flat.map(core.typeof)
   f_debug = api_util.debug_info("custom_root", f, (initial_guess,), {})
-  args_avals = FlatTree.pack(((guess_avals,),{}))
+  args_avals = ft.pack(((guess_avals,),{}))
   f_jaxpr, out_avals = pe.trace_to_jaxpr(f, args_avals, f_debug)
   f_jaxpr, f_consts = pe.separate_consts(f_jaxpr)
 
@@ -114,7 +115,7 @@ def custom_root(f: Callable,
       tangent_solve, (initial_guess, initial_guess), {})
 
 
-  linearize_and_solve_avals = FlatTree.pack(((guess_avals, guess_avals), {}))
+  linearize_and_solve_avals = ft.pack(((guess_avals, guess_avals), {}))
   l_and_s_jaxpr, out_avals = pe.trace_to_jaxpr(
       linearize_and_solve, linearize_and_solve_avals, linearize_and_solve_dbg)
   l_and_s_jaxpr, l_and_s_consts = pe.separate_consts(l_and_s_jaxpr)
@@ -159,8 +160,8 @@ def _root_jvp(const_lengths, jaxprs, primals, tangents):
   linearize_and_solve = partial(
       core.jaxpr_as_fun(jaxprs.l_and_s), *params.l_and_s)
   f_at_solution = lambda *params: f(*params, *solution)
-  _, rhs = ad.jvp(f_at_solution, FlatTree.flatten_list(params.f),
-                  FlatTree.flatten_list(params_dot.f))
+  _, rhs = ad.jvp(f_at_solution, ft.flatten_list(params.f),
+                  ft.flatten_list(params_dot.f))
   solution_dot = _map(
       operator.neg, linearize_and_solve(*solution, *rhs))
   # append aux, create symbolic zero tangents for the aux values
@@ -250,7 +251,7 @@ def custom_linear_solve(
   if transpose_solve is None and symmetric:
     transpose_solve = solve
 
-  b_flat = FlatTree.flatten(b)
+  b_flat = ft.flatten(b)
   b_avals = b_flat.map(core.typeof)
   tree = b_flat.tree
 
@@ -270,7 +271,7 @@ def custom_linear_solve(
   matvec_debug = api_util.debug_info("custom_linear_solve",
                                      matvec, (b,), {})
   # no auxiliary data assumed for matvec
-  args_avals = FlatTree.pack(((b_avals,),{}))
+  args_avals = ft.pack(((b_avals,),{}))
   matvec_jaxpr, out_avals = pe.trace_to_jaxpr(
       _shape_checked(matvec, "matvec", False), args_avals,
       matvec_debug)
@@ -354,8 +355,8 @@ def _tangent_linear_map(func: Callable, params, params_dot,
   """
   assert any(type(p) is not ad_util.Zero for p in params_dot)
   zeros = _map(ad_util.p2tz, x)
-  primals_ft = FlatTree.flatten_list(params + list(x))
-  tangents_ft = FlatTree.flatten_list(params_dot + zeros)
+  primals_ft = ft.flatten_list(params + list(x))
+  tangents_ft = ft.flatten_list(params_dot + zeros)
   _, out_tangent = ad.jvp(func, primals_ft, tangents_ft)
   return list(out_tangent)
