+++ jax/_src/pallas/core.py
@@ -37,6 +37,7 @@ from jax._src import frozen_dict
 from jax._src import hijax
 from jax._src import numpy as jnp
 from jax._src import state
+from jax._src import flattree as ft
 from jax._src import tree_util
 from jax._src import typing as jax_typing
 from jax._src import util
@@ -653,7 +654,7 @@ class BlockSpec:
     with tracing_grid_env(grid, vmapped_dims):
       closed_jaxpr, out_avals = pe.trace_to_jaxpr(
           index_map_func,
-          tree_util.FlatTree(index_map_avals, index_map_tree, False),
+          ft.FlatTree(index_map_avals, index_map_tree, False),
           debug_info)
     unflat_avals = out_avals.unflatten()
 
@@ -1545,7 +1546,7 @@ def core_map(
       fun_args = (scratch_shapes, {})
 
     debug_info = api_util.debug_info("pallas_core_map", f, *fun_args)  # pyrefly: ignore[bad-argument-type]
-    fun_args_refs = tree_util.FlatTree.flatten(fun_args).map(
+    fun_args_refs = ft.flatten(fun_args).map(
         lambda x: x.get_ref_aval())
 
     with (
