+++ jax/_src/pallas/mosaic/pallas_call_registration.py
@@ -28,7 +28,7 @@ from jax._src import frozen_dict
 from jax._src import sharding_impls
 from jax._src import state
 from jax._src import tpu_custom_call
-from jax._src import tree_util
+from jax._src import flattree as ft
 from jax._src.interpreters import mlir
 from jax._src.interpreters import partial_eval as pe
 from jax._src.lib.mlir import ir
@@ -531,7 +531,7 @@ def _rewrite_jaxpr_for_lowering(
   with mpmd.mpmd_map_tracing_context(mesh, all_meshes):
     # TODO(necula): use trace_to_jaxpr_nocache instead
     new_jaxpr, _ = pe.trace_to_jaxpr(
-        new_body, tree_util.FlatTree.flatten_args(*new_in_avals),
+        new_body, ft.flatten_args(*new_in_avals),
         jaxpr.debug_info.with_unknown_names(),
     )
   assert not new_jaxpr.consts
