+++ jax/_src/pallas/mpmd.py
@@ -31,6 +31,7 @@ from jax._src import effects
 from jax._src import hijax
 from jax._src import numpy as jnp
 from jax._src import state
+from jax._src import flattree as ft
 from jax._src import tree_util
 from jax._src import util
 from jax._src.frozen_dict import FrozenDict
@@ -205,7 +206,7 @@ def _mpmd_map_discharge_rule(
         "mpmd_map_discharge", new_body, tracing_avals, {}
     )
     closed_jaxpr, _ = pe.trace_to_jaxpr(
-        new_body, tree_util.FlatTree.flatten_args(*tracing_avals), debug_info)
+        new_body, ft.flatten_args(*tracing_avals), debug_info)
     return closed_jaxpr.jaxpr
 
   for mesh, jaxpr in zip(meshes, jaxprs):
@@ -763,7 +764,7 @@ def _dedup_consts_and_unify_jaxpr_signatures(
     fun_to_trace = make_rewritten_body(jaxpr, consts)
     with mpmd_map_tracing_context(mesh, all_meshes):
       jaxpr, _ = pe.trace_to_jaxpr(
-          fun_to_trace, tree_util.FlatTree.flatten_args(*tracing_avals),
+          fun_to_trace, ft.flatten_args(*tracing_avals),
           debug_info)
     assert not jaxpr.consts, jaxpr.consts
     new_jaxprs.append(jaxpr.jaxpr)
@@ -879,7 +880,7 @@ def _mpmd_map(
         functools.partial(_aval_to_ref_aval, meshes=meshes),
         (kernel_arg_avals, kernel_kwarg_avals),
     )
-    in_avals_ft = tree_util.FlatTree.flatten(unflat_kernel_avals)
+    in_avals_ft = ft.flatten(unflat_kernel_avals)
     flat_kernel_avals = list(in_avals_ft.vals)
 
     jaxprs: list[jax_core.Jaxpr] = []
